ui: refactor log bookmark webview and extract html template
This patch refactors the Log Bookmark webview by extracting its large
inlined HTML, CSS, and JavaScript content into a separate template file
and a dedicated generator class.

1. Resources:
   - Create resources/webview/log-bookmark-template.html to house the UI.

2. Views:
   - Create LogBookmarkHtmlGenerator.ts to handle dynamic content replacement.
   - Refactor LogBookmarkWebviewProvider.ts to delegate HTML generation,
     reducing its size by over 800 lines.

This separation of concerns improves maintainability and makes the
webview provider logic significantly cleaner. It also resolves a bug where
placeholders with whitespace in the template were not correctly replaced.
